# models/fizzbuzz_word.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// FizzbuzzWord fizzbuzz word
//
// swagger:model fizzbuzzWord
type FizzbuzzWord struct {

	// multiple
	// Minimum: 1
	Multiple int32 `json:"multiple,omitempty"`

	// word
	Word string `json:"word,omitempty"`
}

// Validate validates this fizzbuzz word
func (m *FizzbuzzWord) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateMultiple(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *FizzbuzzWord) validateMultiple(formats strfmt.Registry) error {
	if swag.IsZero(m.Multiple) { // not required
		return nil
	}

	if err := validate.MinimumInt("multiple", "body", int64(m.Multiple), 1, false); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this fizzbuzz word based on context it is used
func (m *FizzbuzzWord) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *FizzbuzzWord) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *FizzbuzzWord) UnmarshalBinary(b []byte) error {
	var res FizzbuzzWord
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}