# restapi/operations/fizzbuzz_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package operations

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"

	"github.com/pastequo/services.fizzbuzz/models"
)

// FizzbuzzOKCode is the HTTP code returned for type FizzbuzzOK
const FizzbuzzOKCode int = 200

/*FizzbuzzOK Successfully applied.

swagger:response fizzbuzzOK
*/
type FizzbuzzOK struct {

	/*
	  In: Body
	*/
	Payload string `json:"body,omitempty"`
}

// NewFizzbuzzOK creates FizzbuzzOK with default headers values
func NewFizzbuzzOK() *FizzbuzzOK {
	return &FizzbuzzOK{}
}

// WithPayload adds the payload to the fizzbuzz o k response
func (o *FizzbuzzOK) WithPayload(payload string) *FizzbuzzOK {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the fizzbuzz o k response
func (o *FizzbuzzOK) SetPayload(payload string) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *FizzbuzzOK) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {
	rw.WriteHeader(200)
	payload := o.Payload
	if err := producer.Produce(rw, payload); err != nil {
		panic(err) // let the recovery middleware deal with this
	}
}

// FizzbuzzNoContentCode is the HTTP code returned for type FizzbuzzNoContent
const FizzbuzzNoContentCode int = 204

/*FizzbuzzNoContent Successfull but empty.

swagger:response fizzbuzzNoContent
*/
type FizzbuzzNoContent struct{}

// NewFizzbuzzNoContent creates FizzbuzzNoContent with default headers values
func NewFizzbuzzNoContent() *FizzbuzzNoContent {
	return &FizzbuzzNoContent{}
}

// WriteResponse to the client
func (o *FizzbuzzNoContent) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {
	rw.Header().Del(runtime.HeaderContentType) // Remove Content-Type on empty responses

	rw.WriteHeader(204)
}

// FizzbuzzBadRequestCode is the HTTP code returned for type FizzbuzzBadRequest
const FizzbuzzBadRequestCode int = 400

/*FizzbuzzBadRequest Bad request.

swagger:response fizzbuzzBadRequest
*/
type FizzbuzzBadRequest struct {

	/*
	  In: Body
	*/
	Payload *models.ErrorMsg `json:"body,omitempty"`
}

// NewFizzbuzzBadRequest creates FizzbuzzBadRequest with default headers values
func NewFizzbuzzBadRequest() *FizzbuzzBadRequest {
	return &FizzbuzzBadRequest{}
}

// WithPayload adds the payload to the fizzbuzz bad request response
func (o *FizzbuzzBadRequest) WithPayload(payload *models.ErrorMsg) *FizzbuzzBadRequest {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the fizzbuzz bad request response
func (o *FizzbuzzBadRequest) SetPayload(payload *models.ErrorMsg) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *FizzbuzzBadRequest) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {
	rw.WriteHeader(400)
	if o.Payload != nil {
		payload := o.Payload
		if err := producer.Produce(rw, payload); err != nil {
			panic(err) // let the recovery middleware deal with this
		}
	}
}

// FizzbuzzInternalServerErrorCode is the HTTP code returned for type FizzbuzzInternalServerError
const FizzbuzzInternalServerErrorCode int = 500

/*FizzbuzzInternalServerError Internal error.

swagger:response fizzbuzzInternalServerError
*/
type FizzbuzzInternalServerError struct{}

// NewFizzbuzzInternalServerError creates FizzbuzzInternalServerError with default headers values
func NewFizzbuzzInternalServerError() *FizzbuzzInternalServerError {
	return &FizzbuzzInternalServerError{}
}

// WriteResponse to the client
func (o *FizzbuzzInternalServerError) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {
	rw.Header().Del(runtime.HeaderContentType) // Remove Content-Type on empty responses

	rw.WriteHeader(500)
}

// FizzbuzzServiceUnavailableCode is the HTTP code returned for type FizzbuzzServiceUnavailable
const FizzbuzzServiceUnavailableCode int = 503

/*FizzbuzzServiceUnavailable Not available.

swagger:response fizzbuzzServiceUnavailable
*/
type FizzbuzzServiceUnavailable struct{}

// NewFizzbuzzServiceUnavailable creates FizzbuzzServiceUnavailable with default headers values
func NewFizzbuzzServiceUnavailable() *FizzbuzzServiceUnavailable {
	return &FizzbuzzServiceUnavailable{}
}

// WriteResponse to the client
func (o *FizzbuzzServiceUnavailable) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {
	rw.Header().Del(runtime.HeaderContentType) // Remove Content-Type on empty responses

	rw.WriteHeader(503)
}